+++ src/main/java/br/com/dio/board/Main.java
@@ -1,4 +1,4 @@
-package br.com.dio.dio_design_pattern_project;
+package br.com.dio.board;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
