+++ src/main/java/br/com/dio/board/repository/IBoardRepository.java
@@ -1,9 +1,9 @@
-package br.com.dio.dio_design_pattern_project.repository;
+package br.com.dio.board.repository;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import br.com.dio.dio_design_pattern_project.model.Board;
+import br.com.dio.board.model.Board;
 
 @Repository
 public interface IBoardRepository extends CrudRepository<Board, Long> {
