+++ src/main/java/br/com/dio/board/repository/ICardRepository.java
@@ -1,9 +1,9 @@
-package br.com.dio.dio_design_pattern_project.repository;
+package br.com.dio.board.repository;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import br.com.dio.dio_design_pattern_project.model.Card;
+import br.com.dio.board.model.Card;
 
 @Repository
 public interface ICardRepository extends CrudRepository<Card, Long> {
