+++ src/main/java/br/com/dio/board/repository/IColumnRepository.java
@@ -1,9 +1,9 @@
-package br.com.dio.dio_design_pattern_project.repository;
+package br.com.dio.board.repository;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import br.com.dio.dio_design_pattern_project.model.Column;
+import br.com.dio.board.model.Column;
 
 @Repository
 public interface IColumnRepository extends CrudRepository<Column, Long> {
