+++ src/main/java/br/com/dio/board/service/BoardService.java
@@ -1,6 +1,6 @@
-package br.com.dio.dio_design_pattern_project.service;
+package br.com.dio.board.service;
 
-import br.com.dio.dio_design_pattern_project.model.Board;
+import br.com.dio.board.model.Board;
 
 public interface BoardService {
 	Iterable<Board> getAll();
